refactored router and components to separated concerns
