+++ public/index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Boilerplate</title>
+  <title>Expensify</title>
   <link rel="icon" type="image/x-icon" href="/favicon.ico">
 </head>
 <body>
