+++ src/components/AddExpensePage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function AddExpensePage() {
+  return (
+    <div>
+      This is from Add Expense Compy
+    </div>
+  );
+}
