+++ src/components/EditExpensePage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function EditExpensePage() {
+  return (
+    <div>
+      This is from Edit Expense Compy
+    </div>
+  );
+}
+
