+++ src/components/ExpenseDashboardPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ExpenseDashboardPage() {
+  return(
+    <div>
+      This is from Dashboard Compy
+    </div>
+  );
+}
