+++ src/components/Header.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+// Alternate ES6 Stateless Functional Component
+const Header = () => (
+  <header>
+    <h1>Expensify</h1>
+    <NavLink to="/" activeClassName="is-active" exact> Dashboard </NavLink>
+    <NavLink to="/create" activeClassName="is-active"> Create Expense </NavLink>
+    <NavLink to="/edit" activeClassName="is-active"> Edit Expense </NavLink>
+    <NavLink to="/help" activeClassName="is-active"> Help </NavLink>
+  </header>
+);
+
+export default Header
