+++ src/components/HelpPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function HelpPage() {
+  return (
+    <div>
+      This is from Help Compy
+    </div>
+  );
+}
+
