+++ src/routers/AppRouter.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { Route, Switch, } from 'react-router-dom';
+import Header from '../components/Header';
+import ExpenseDashboardPage from '../components/ExpenseDashboardPage'
+import AddExpensePage from '../components/AddExpensePage';
+import EditExpensePage from '../components/EditExpensePage';
+import HelpPage from '../components/HelpPage';
+import NotFoundPage from '../components/NotFoundPage';
+
+export default function AppRouter() {
+  return (
+    <div className="AppRouter">
+      <Header />
+      <Switch>
+        <Route exact path="/" component={ExpenseDashboardPage} />
+        <Route path="/create" component={AddExpensePage} />
+        <Route path="/edit" component={EditExpensePage} />
+        <Route path="/help" component={HelpPage} />
+        <Route component={NotFoundPage} />
+      </Switch>
+    </div>
+  );
+}
